+++ packages/paste-website/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.19](https://github.com/twilio-labs/paste/compare/@twilio-paste/website@2.14.18...@twilio-paste/website@2.14.19) (2020-11-25)
+
+**Note:** Version bump only for package @twilio-paste/website
+
+
+
+
+
 ## [2.14.18](https://github.com/twilio-labs/paste/compare/@twilio-paste/website@2.14.17...@twilio-paste/website@2.14.18) (2020-11-19)
 
 **Note:** Version bump only for package @twilio-paste/website